lbbs: Fix gcc 11.3.0 compiler warnings.
Fixes compiler warnings raised by gcc 11.3.0,
not previously caught with gcc 10.2.1 or 8.3.0,
since Ubuntu comes with a newer gcc than Debian.

The main warning addressed here is -Wstack-protector.
Several usages of strdupa are converted to use fixed
or dynamically allocated buffers, some of which were
possibly unsafe if the buffer size was determined via
any kind of user input, others of which were okay but
converted anyways. A few usages remain and a pragma
is used to suppress the warnings for these exceptions.