lbbs: Fix SIGWINCH race condition.
A race condition turned out to be possible whereby
the parent would send the child its initial SIGWINCH
before the child had removed the parent's SIGWINCH
handler. Consequently, the child would execute the handler
and deadlock on a bbs_log function that it wasn't supposed
to be executing in the first place.

Because this happened after fork() but prior to exec(),
the child's signal handlers remained the same as the parent's,
so if a user pressed ^C to send a SIGINT to the child, it
would end up shutting down the entire BBS.

To fix this, we now check the PID in the SIGWINCH signal handler
and bail early if it's not the parent PID, since this could
happen during the race condition described above. The same thing
is done for the SIGINT handler, even though this was not necessary
to fix this issue.