socket.c: Log IP address on accept()
Get the IP address after accept() is called
so that we can log it immediately. Although
we log the IP address when the node handler
starts, for the generic TCP listener, if login
or registration does not proceed, the address
never gets logged. Similar to what the SSH
network module does, this ensures that the IP
address is logged immediately upon a connection.