pty: Minor usability improvements.
TDDs will send CR and LF, but in separate
reads, which will trigger the RLogin workaround,
so only do this if the protocol is actually RLogin,
to avoid turning CR LF into CR LF LF erroneously.

Also print a message to the sysop if the node being
spied on has disconnected.

Backtrace addresses can be as wide as 14 characters,
so I've also increased the room allocated to that.