lbbs: Add TTY/TDD support.
Adds integrated support for TDDs.
There are several parts to this commit:

First, revert code in pty.c that restricted CR LF
emulation to RLogin, since this actually applies
to Telnet as well. To make it service agnostic,
we always emulate and ignore subsequent LFs that
follows immediately afterwards.

Secondly, a function is added that strips
ANSI escape sequences from a buffer. This is needed
since TDDs aren't ANSI-capable terminals, and we
should avoid sending them stuff they don't want.

Finally, some node properties are added to make use
of this functionality, and the Telnet driver is extended
to also double as the TTY driver, since it's basically
the same thing. The difference is if we receive a connection
on the TTY/TDD port, we assume it's a TDD and disable ANSI
for the node and set the dimensions to 24x1, a TDD's window
dimensions.

At present, this slightly improves the experience for TDDs
since ANSI escape sequences aren't sent to them. However,
there is still more work to be done, such as optimizing
prompts, menus, etc. for small screens like a TDD's.