lbbs: Continue enhancing TDD support.
Adds some further optimizations for TDD devices, including correcting
the terminal dimensions that are set manually. I thought TDDs had
24 columns, but they actually seem to have only 20.

Additionally, there used to be an assertion in the node I/O functions,
e.g. bbs_read, bbs_write to assert the slave file descriptor
was not -1. However, after these assertions have triggered
numerous times, it has become clear this is not an appropriate
assertion, since we can't always guarantee that it won't be -1
by the time the function is called, even with a judicious use
of NEG_RETURN macros all over the codebase. Therefore, we now
just warn and return -1, which is more appropriate, as it will
cause the node to exit cleanly in due course, and if we see a
lot of these, then it can still be investigated. The main effect
is that if the BBS is started with the -g option, it will no longer
abort on these assertions, which previously happened frequently.

A change to the softmodem that was made was to automatically
send 2 spaces of output whenever output starts again (from
idle carrier). This is to prevent clipping of output from the BBS.
Rather than having the softmodem send LTRS in case the TDD is in FIGS
mode, we send some spaces upon a new TDD connection, first thing,
to ensure the actual beginning of printout is not corrupted.

One issue that remains is that, at least my TDD, seems to send
CR LF CR when I hit RETURN, which gets interpreted as 2 newlines,
basically. Furthermore, some symbols don't seem to get sent,
largely because not all symbols have a Baudot encoding since
Baudot is 5-bit and ASCII is 7-bit. Some kind of translation
scheme will be needed in contexts that require them.
This will still need to be figured out.