pty: Sanely handle CR LF CR from TDDs.
TDDs (at least mine) send CR LF CR when you press
RETURN, rather than CR LF as would be expected.
As a result, this has hitherto been interpreted
as 2 "ENTER"s. To fix this, some logic is added
to the pseudoterminal relay to check for this.
Although this does not explicitly require a TDD,
the logic only applies when CR, LF, and CR are
received sequentially and separately, which would
generally only happen with something as slow as TDD.

Some more TDD output enhancements are also present
in this commit.