lbbs: Fix issues with TDD I/O.
This makes two major improvements for TDD I/O.

The first is to add a mechanism that allows
characters from nodes to be replaced with others
in the input stream. This is only done when a single
character is received, currently, by pty.c, so it
really only applies to TDDs, since it's not needed
for other terminals (which presumably have a full keyboard).
This allows symbols that a TDD can encode in Baudot
to temporarily emulate the actual symbol desired for input.

A related issue was that characters were being echoed back
to the TDD. Because TDDs essentially always have local echo on,
and this cannot be disabled (since for the use case of a TDD,
this generally would not make sense), we compensate for this
by overriding any echo sets to ensure that echo always remains
disabled for TDDs. This prevents output from interrupting
input on the TDD, which was a real problem since TDDs use
a half duplex channel.