door_irc: Add IRC client module.
Adds an IRC client door module, that
uses (and links to) the LIRC library
recently written for this purpose.

Currently, each door access allows
a user to participate in one IRC
channel using a specified client. The
same client can be used for as many
users and channels as are accessible
through that client (same IRC server).