menus: Further optimize TDD experience, misc. enhancements.
Rather than printing all menu options by default, we now
skip this by default for TDDs, and users can press SPACE
to get a full menu listing at any point. This will draw
the full menu.

A couple of other things are now skipped for TDDs as well.

Finally, I noticed that if a timeout warning happened
on a menu, the current line was not getting cleared. This
is now fixed by clearing the line regardless of the node
buffered setting.