nodes: Fix socket shutdown and spy closure.
Call shutdown() on the node's socket file descriptor
so that poll() is interrupted immediately and the
node can exit immediately.

Also fixes a remaining instance of writing to
STDOUT for when sysop spying is terminated. This
is fixed by storing the sysop's STDOUT file descriptor
on the node and writing to that instead.