net_sftp: Add SFTP server module.
Adds a network module that provides a basic SFTP server.
I ended up making this completely separate from the SSH
server module, although it does use the same underlying
libssh library. As a result, SFTP and SSH cannot currently
run on the same port. We could work on fixing that in
the future by using a single listener in the SSH module.
One reason for doing it this way for now is the architecture
of the two modules is completely different: the SSH module
is event-driven whereas the SFTP module is written using
the older libssh interface.