logger: Add logging for remote sysop consoles.
Although commit bf486e104a58e1268444683a6ea6f8fb147d97cf
previously added support for remote sysop consoles,
this was purely a command console interface that did
not output log messages from the system, making the
remote console experience significantly inferior
to using the foreground console.

This adds support for logging to remote sysop consoles
in much the same way that we log to the foreground
console, if it is available. With this change, the
experience for both the foreground and remote consoles
should be more or less identical.