utils: Add backspace processing function.
Some terminal emulators e.g. SyncTERM don't seem
to properly handle backspaces even when the terminal
is in canonical mode. As a result, backspaces work
fine in programs like PuTTY/KiTTY but not in SyncTERM.

This isn't really our fault (at least, I don't think it is),
but in an attempt to improve the user experience and consistency,
we now actually process these control characters, rather than
allowing them to be part of the processed input stream.