mod_sysop: Fix lingering history after ESC.
Currently, if a user uses arrow keys to access
a command in history and then presses ESC,
the history pointer lingers behind, so if
a user presses ENTER, the command that was
last accessed from history will execute.
This is obviously incorrect, but also dangerous
since there's no indication or confirmation.
This is fixed by properly clearing the command
reference when ESC is pressed to fully reset.