net_finger: Add Finger server.
Adds a basic RFC 1288 Finger server module.
Currently very rudimentary, but so too is the Finger protocol itself.
This is noted in the module and the sample config, but Finger is not
a secure protocol, in any way, so we don't want to provide more
information than would be prudent using this protocol.
Query forwarding is not supported. Vending machines are also not supported.
The listing still needs to be expanded; currently, it just
prints out the user name and user ID for a successful user query.