net_http: Improve Basic Authentication performance.
This significantly improves the performance of Basic
Authentication in the HTTP web server. Previously,
we were authenticating normally on every request that
used Basic Authentication (which is all of them, if needed),
which is very slow as crypt_r is slow (which is not bad,
but causes a noticable delay in this case).

To improve this, we now hash passwords using SHA256
and store them in memory along with the username in a
small cache. This way, on successive web requests after
the initial authentication, we can SHA256 the password
and compare and, on a match, reauthenticate the user
without having to call crypt_r or other crypto functions.

Obviously, this is not quite as secure, since SHA256
is a one-way hash that will return the same hash for
the same password. The risk is not as great since we
are not persisting passwords using this mechanism, and
the hashes are stored only in volatile memory. Importantly,
passwords cannot be reconstructed from the hashes.

To prevent caching from building up, we only keep a limited
number of hashed passwords in memory.