mod_auth_mysql: Add datetime binding, dynamic string result allocation
Increases the flexibility of the MySQL auth module by
allowing dates to be passed in as struct tm's for
columns that have a datetime type.

Consequent of this, two fields that were previously ignored
on registration are now inserted into the table (and the
schema has been updated appropriately).

Also simplifies binding results by allowing string
result fields to be dynamically allocated based on
their actual length, rather than trying to dynamically
allocate a large enough buffer in advance (which may
potentially be too small, and if not, might waste memory).