mod_mysql: Put MySQL interface in separate module.
This moves the MySQL/MariaDB interface into a separate
module that could be used by multiple MySQL-related modules.

To enable this, modules are now able to export global symbols.
In order to enforce the proper loading order (modules that export
global symbols required by other dependent modules must load
before they do), the preload directive is now added, which can
be used to ensure dependencies load first (in addition to exporting
global symbols). This is necessary since we pass RTLD_NOW to
dlopen (which is definitely a good practice).

A couple mechanisms are also added to ensure that dependencies
can't be unloaded while things that are dependent on them are
still loaded.