socket.c: Add dual-socket listener function.
Since it's common for some services to listen
on two ports, a plain text port and a secure port,
e.g. HTTP/HTTPS, IRC/IRCS, etc., adds a function
to socket.c that can listen on both ports,
to simplify listening setup for modules that need
to listen on 2 ports but use the same handler
for both sockets.