mod_mysql: Fill in incomplete tm struct for date results.
The MYSQL_TIME object doesn't contain any fields corresponding
to things like day of the week, which means if we merely copy
and translate the fields that do map over, then using this
in other operations like strftime will always result in
printing out "Sunday" as day of the week.

To fix this globally, we now convert to the epoch timestamp
and then convert immediately back to a tm struct, letting
localtime do the work of figuring out these other fields and
then copying these over into the result struct. This way,
these fields are available if they are needed.