module.c: Fix leak for modules exporting global symbols.
Fixes a memory leak for modules that export global symbols,
since the way we handle this is by opening modules with
RTLD_LOCAL first, checking if the module needs to export
global symbols and, if so, closing and reopening with
RTLD_GLOBAL instead. However, we were not actually freeing
the module, leading to a memory leak. This is now fixed.