curl: Fix pointer oddity and response handling for non-HTTP
This addresses two issues in the cURL interface. First off,
cURL can be used for protocols other than HTTP, so this
checks the response code using HTTP semantics only if
the request was actually an HTTP/HTTPS request.
Secondly, there can be an issue with response->str
not being valid when returning to curl_common_run,
but duplicating the pointer in the write callback
and using that suffices as a workaround. I'm not exactly
sure why, so this has "hack" written all over it, but it
does work and prevents dangling pointers, things being
NULL that shouldn't be, memory leaks, and segfaults, so
I'm adding this workaround for now. Ideally though, we
should figure out why str isn't working properly.