net_sftp: Increase portability with older libssh.
Older versions of libssh, prior to commit
cc536377f9711d9883678efe4fcf4cb6449c3b1a
do not have sftp_server_free. Furthermore,
there is no easy way to tell from libssh
header files if this function is even available.
Therefore, determine in the Makefile if the
libssh library exports this symbol, and if not,
duplicate the necessary code in the module
so that compilation and linking succeed.