Makefile: Improve valgrind execution portability.
valgrind prior to version 3.15.0 did not have
the -s or --show-error-list options.
Providing these will cause valgrind to abort,
and so using the valgrind targets has hitherto
failed on systems with older valgrind versions.
The workaround was manually running the command
the target tried to run, without the --show-error-list
option.

This adds version checking into a separate make
target itself, so that we can automatically
run the appropriate command based on the install
valgrind version, eliminating this clunkiness.