tls.c: Properly relay TLS traffic to/from sockets.
SSL_get_rfd and SSL_get_wfd merely return the file
descriptor passed into SSL_set_fd in the first place,
making these functions rather useless for our purposes,
since they are the raw socket itself, not the decrypted
side of it. As a result, HTTPS did not work at all
since we were interacting with the socket directly.

To work around this, we now spawn another thread that
relays traffic from encrypted connections. Since this
will mostly be buffered anyways, for things like HTTP
requests and IRC messages, etc., a single thread is used
for efficiency.

This allows us to continue with the model of abstraction
that was desired, using a file descriptor (or two, rather,
one for reading and one for writing, since we're using
pipes internally and not sockets) so that applications
do not have to be concerned with whether the connection
itself is encrypted or not. This allows the presentation
layer (TLS) to be effectively separated from the application layer.

Fixes #1