net_irc, tls.c: Crash and loop fixes.
If a user isn't logged in, we still call welcome,
so don't attempt to dereference user if it's NULL.

Since tls.c relays between file descriptors using
read/write, we need to close the application side
of the relay if the socket side closes, or we'll
get stuck in an infinite loop as read will return 0
forever but the node will never see this. This is
fixed so disconnects pass propagate properly now.