mod_discord: Add Discord/IRC relay.
Adds a module to provide a simple relay
between IRC channels and Discord channels.

mod_discord requires being linked with
libdiscord, the Concord Discord library (written in C).
This is not compiled in the Makefile, but is compiled
and installed as part of the prereqs, since it only
needs to be done once.