net_irc, mod_discord: Add NAMES, WHO, WHOIS integration.
Adds a callback mechanism for relays to provide responses
for NAMES, WHO, and WHOIS replies on IRC. This allows
Discord members, for example, to appear in IRC, and for
IRC members to perform simple queries on them. This functionality
may continue to be extended in the future.

Currently, there are some bugs either with libdiscord or the usage
of the library such that a couple things don't work correctly.
First off, we are unable to get the user presences when the module
loads (only updates thereafter), so we don't know immediately
who is actually in the channel or not. Secondly, we don't know
what channels a Discord user is in, which is necessary since the
member list corresponds to the entire guild, but some channels
may not contain all members. For now, we presume all members
in the guild are in all channels, just so that it works, but this
is not a reasonable assumption and will need to be fixed.

Finally, there are some multithreading usage issues with libdiscord
itself that have been reported and should be fixed soon. The crux
of this issue is that we need to be able to stop discord_run from
a thread other than the thread currently running it. However,
this can either result in blocking forever or segfaulting since
this is not multithread safe. So, unloading the module at present
is currently not stable, but hopefully this should be resolved soon
with the addition of a libdiscord API to set the shutting down flag
in the same way as its signal handler option does now.

Also fixes the server creation time by using the module
load time instead of the BBS start time, which is more appropriate.