tls.c, mod_discord: Minor crash fixes.
Based on some crashes that occured in these two
places recently, adds mitigations for these.

In tls.c, it appears that we can try to read
from an SSL* that has been freed, so hold
a RDLOCK when reading/writing so that SSL*'s
can't disappear while we're doing that.

In mod_discord, there was a crash observed on
this line, presumably because there was no
event, so emit an error and abort if author
is NULL for whatever reason.