door_irc: Fix timeouts causing IRC client door to exit.
An issue existed where taking more than 30 seconds to
compose a message would result in the IRC client exiting
abruptly. This was because the node was buffered, and poll
would return 0 after 30 seconds since the message had not
yet been finished.

This affected all nodes, but was particularly noticable
for TDD connections, since it can easily take 30 seconds
to type a message at 45.45/50 baud.

To fix  this, we now continue rather than break if poll
returns 0, and simply time out the message, rather than
exit entirely, and the timeout is also increased from 30
seconds to 3 minutes, since the former is inappropriately low.