thread.c: Allow getting TID/LWP by pthread_t.
POSIX threads (pthreads) provides no mechanism
of getting the TID/LWP from a pthread_t, but
since we store both of these for all registered
threads, we can easily determine this.

To make debugging easier, we now print out the
thread ID/LWP for each node, if one exists,
when either all nodes or a single node is dumped.