mod_auth_mysql: Don't allow spaces in usernames.
Some services integrated into the BBS, e.g. IRC,
don't allow spaces in usernames, so for maximal
compatibility, don't allow registrations for usernames
containing spaces, to avoid issues down the line.

In addition to this, at certain points in the core,
if we detect an existing username with spaces, we
make some noise to alert the sysop to this incompatibility.