mod_discord: Fix presences on startup, channel membership.
Major changes to address the previous major issues of not
having presence information at load time and not knowing
what users were members of what channels.

We are now able to get presence information in the guild
member chunks due to a libdiscord bug fix (currently on
the dev branch) as well as setting .presence = true.
However, the presence count is not guaranteed to be (and
seems to be typically less than) the member count available
in each chunk, so we need to retry these requests, filtering
on members for whom we previously failed to get presence info,
until we have gotten all of them (or at least, until we stop
getting more presences, since some can repeatedly fail).

Due to a likely libdiscord JSON serialization issue caused
by buffer truncation if we retry with too many user IDs in
the filter, we currently limit retries to 25 users at a time
to work around this issue. This limit can be removed later
if this issue is fixed.

The other issue was not being able to determine what members
are in what channels, or what channels contain what members,
said another way. I can't currently find a good way to do this
with the Discord API and/or libdiscord. The workaround is to use
channel permissions as a crude proxy for membership. By default,
channels are open to everyone, but they can be made private with
select users or roles allowed to join. This method works well,
although we need to explicitly check for administrators and owner
as well, and for some reason we are unable to get the guild owner's
user ID at the moment. Once this is fixed, this should be fully
working. Ironically, the guild owner is currently excluded from
private channel memberships.

With this change, the nicklist integration now works sensibly.
There are a few remaining outstanding issues that affect usability
of this module:
- Unable to get guild owner ID
- Segfaults are still probable on module unload due to race conditions

Additionally, if the serialization buffer issue is fixed, we should
remove the limit of 25 users per retry to reduce the number of requests.