net_irc: Resolve IP addresses to hostnames.
IRC servers are supposed to resolve IP addresses
to hostnames on client connects. Previously, we
were just taking the IP address and running with it,
but now we actually do a reverse lookup for the hostname.
This information is not meaningfully used, other than
being sent in a NOTICE message to the client, since
all clients are automatically cloaked when authenticated.