net_irc, mod_relay_irc: Fix WHO replies.
Fix WHO reply formatting so that IRC clients properly
display the away status of users in the nicklist.
One major issue was we were using send_numeric instead
of send_numeric2, and as such, an extra colon was being
inserted erroneously, causing clients to misparse the message.
This is now fixed, and clients now properly show the away
status of both local network users as well as users on
other IRC networks.