mod_discord: Fix crashes on unload, away status.
Previously, mod_discord would typically cause
a segfault when unloaded, due to race conditions
and multithread unsafeness in concord. This has
been addressed upstream in this commit:
https://github.com/Cogmasters/concord/commit/906f7cae5526e8782395e47233aa2a3418e5e14b

Using the added API, the module now unloads gracefully.

Additionally, WHO responses are fixed the same way
as in net_irc (remove erroneous colon) and now users
that are offline or idle now show up as "away" on IRC.

One remaining issue is that we are currently unable to
get the presence data of all guild members initially,
and retries do not get additional members, only the
same member presences again. This is still being investigated.