mod_discord: Fix relaying of system messages.
Previously, system messages through an IRC relay
(as opposed to directly from the local IRC server)
were not being relayed to Discord.

irc_relay_raw_send is now modified to support
this in the same way that irc_relay_send was in
commit 0e033651911c05b6ee4bc25d28df731d2c46275b

Additionally, mod_discord is updated to parse these
raw IRC messages and format them in the same way that
system messages from the native IRC server are formatted.