mod_relay_irc, mod_discord: Fix CTCP relaying and add startupjoinignore.
Fixes relaying of CTCP messages from IRC, both from the local network
to other IRC networks and to Discord.

Also adds a setting, startupjoinignore, that can be used to reduce
annoyance in IRC and Discord channels by suppressing the flood of
IRC client reconnects when the local IRC server first starts (presumably
the same time that the relay module is starting as well).