module.c: Automatically unload/reload dependents as needed.
Since some modules can depend on other modules, these dependencies
are now codified into the modules themselves, rather than just
being noted in the sample modules.conf file.

As a result, one thing we now do is attempt to automatically
unload any modules that may be dependent on a module that we want
to unload. Previously, the sysop would have to manually unload
modules that dependend on the target module first, and then unload
that module. In the case of a reload, the unloaded modules would then
also have to be manually loaded again. Although there is a slight
risk that a user may not intend to auto-unload dependent modules,
generally this results in a much better user experience to handle
this automatically.

Note that this commit only handles dependents for modules that are
already running. It doesn't do anything at startup with regards to
handling loading of dependencies of a module.