mod_relay_irc: Add rudimentary loop avoidance.
This has only been observed once, but on one system
a messaging loop happened near startup due to the JOIN
message propagating back and forth forever. One attribute
of this phenomenon is that because we prefix usernames with
the client name, the nickname will grow longer and longer
initially, until eventually it fills the entire buffer
available. Therefore, we can crudely detect such a loop
in progress by simply checking if the nickname is using
the entire buffer or not, and this is a pretty good indication
that a loop happened. This doesn't address how loops might
start in the first place (possibly even due to configuration
error), but it at least stops some loops almost as soon as
they get started.