mod_relay_irc: Prevent crash on nick changes from IRC clients.
If the IRC relay module processes a message from an IRC client
and channel is NULL, it will crash. Because there's no clear
way to determine if we should be relaying at all if we don't
have a channel with which we can look up a channel mapping,
just drop such messages for now.