module.c: Automatically preload dependencies.
Now that commit 582ed86712352c58be3699b60dd17dbea0837633
allows modules to codify their dependencies, we can use
these to automatically load any dependencies a module has
before actually loading the module for real. This improves
the user experience since proper module loading no longer
requires all the module dependencies to be explicitly
preloaded in modules.conf. This will be done automatically,
as long as the dependencies are not noloaded (in which case,
a warning will be emitted, since something will be unable
to load as a result).