config: Add function to set directory path from config setting.
Loading a directory path into a string buffer is a common enough
task that this adds a function to do that to config.c.
This eliminates duplicated code for detecting a trailing slash
in a buffer and eliminating it if present, as well as warning
if the directory does not exist or is not readable.