user.c: Add API to get user ID by username.
Adds a function that can look up a user ID by username.
This is encapsulated into a common function since the
current implementation is not optimal, and if optimized
in the future, all users will benefit from using the
common implementation. Additionally, if trying to do this
manually, callers might forget to call bbs_user_destroy
and leak memory.