utils: Add function to read single line.
Adds a function to read a single line from a file descriptor.
Despite sounding simple, this is actually somewhat complicated
if we want to account for the possibility of read returning
partial or multiple lines (which can happen). Doing this manually
in modules that might require the high level of correctness required
is likely to be messy and duplicative, so this is a single common
implementation that can be used throughout the tree, e.g. for reading
from network sockets.