utils: Fix bbs_fd_readline not advancing buffer.
In the simple (common) case where we read only one
chunk of input (e.g. one line) from a file descriptor
in one round, the buffer was not properly advanced
and we would keep reading from the beginning while
additional input would get queued up behind it.
This is fixed by setting our current position properly
if there are no leftovers.

Some additional tests are added to the unit test for
this function which captures this condition.

A few test macros are also moved into the header file
so that tests can be declared outside of mod_test.