mod_sysop, pty: Fix file descriptor leak.
The actual socket file descriptor that mod_sysop
would obtain from accept() was not being closed,
so file descriptors were piling up from remote
sysop sessions. This is fixed by closing not
only the other file descriptor in the PTY master
relay thread, but also the current file descriptor,
since nobody else is going to close it for us,
in this scenario.