net_irc: Fix erroneous client disconnects.
IRC clients were periodically (seemingly at random)
being disconnected from the IRC server. The issue
is that bbs_fd_poll_read will return 0 if either poll
or read returns 0, and the poll was timing out due
to an insufficient poll interval (strictly PING_TIME).
This could be problematic as the ping thread may actually
take just slightly longer than this to return to the same
client on the next round, which could result in clients
exceeding the timeout since they won't yet have sent
a PONG response.

This timeout is now increased, to twice that,
which should be more than sufficient.