net_smtp, mod_mail: Add SMTP server.
This adds a functional SMTP server to the BBS.
The SMTP server is both a Mail Transfer Agent
(MTA) as well as a Message Submission Agent (MSA).
The MSA accepts outgoing mail from local users
to other local mailboxes or external addresses.
The MTA accepts external mail for local users and
relays outgoing mail from local users to other
SMTP servers. A configurable mail queue also
allows for retries if the server is unable to
deliver a message to another SMTP MTA on the
first attempt.

The SMTP server is functional, although currently
only useful for sending external mail in practice
as there is no user-friendly way to check messages
(e.g. POP3/IMAP4) yet. The messages are merely
saved to disk, using the maildir file/directory format.

Some generic mail functions that are likely to
be useful for POP3/IMAP4 are added in a separate
mod_mail resource module, so that these can be
used between all email modules.