module.c: Autokick nodes started by a module when reloading it.
Although we currently automatically unload any modules dependent
on a module before attempting to unload it, this doesn't account
for any nodes that may have been created by that module, and
as such, unloading will fail as long as those nodes are active.
For some modules, such as net_irc, this can be tricky as many
IRC clients will attempt to immediately reconnect if the connection
drops, so even if you manually kick these nodes and then reload,
they might have rejoined already.

This adds a mechanism to automatically kick nodes that were spawned
by a particular module. Note that this does NOT include kicking
modules that may be in use by any nodes (e.g. doors). In particular,
this really only applies to network modules, since that's how all
connections arrive to the BBS. This increases the chance of an unload
succeeding.

Additionally, behavior is modified such that a reload will autokick
nodes in this manner, but an unload will not and therefore will fail
if there are active nodes spawned by the module.