net_smtp: Add configurable relay restrictions.
Adds settings that enforce a minimum privilege
level for receiving external email as well
as sending external email.

This adds an API to user.c to accomplish this,
but the implementation is truly horrible since
it's linear time instead of constant time. A new
API likely needs to be added to auth.c (with
mod_auth_mysql integration) to resolve this.

Some fixes are also included in mod_mysql,
notably fixing permissions for directory creation,
a truncation warning in mailbox creation, and
ensuring we create default directories (e.g. Sent,
Trash, etc.) with proper maildir++ format names.