net_imap: Add IMAP4 server.
This adds a module to provide an IMAP4 server
for mail clients to view messages on the server.
This is still incomplete (the entire RFC is not
yet implemented), but basic operations akin to
what you can do with POP3, i.e. retrieving messages
in the INBOX, marking them as read/unread, etc.
are currently functional, so users can at least
manage their mail. Other operations will
continue to be added.