test.c: Fix percent tests passed calculation error.
The percent of tests passed calculation was erroneously
being calculated based on the first test failure, since
we were checking the return value OR'ed with all previous
return values, rather than simply the return value, to
determine if the test succeeded. This meant that any test
failure would cause all subsequent tests to be calculated
as failed, regardless of whether they passed or not.
This is now fixed.