net_nntp: Add newsgroup server.
Adds a basic newsgroup server to the BBS.
Currently, it is an isolated server that
allows for articles to be posted and downloaded
from the server only. However, in the future
it can support NNSP between servers for synchronizing
articles. Many commands are also still missing,
so this is really just a minimally viable module
at the moment.

Closes #3