module.c: Decrement dependency use counts on shutdown.
Module use counts of dependencies are decremented if
they are unloaded manually during runtime, but not when
they are unloaded during shutdown, which caused modules
that were dependencies of other modules to not properly
unload during shutdown, since their use count would never
fall to 0. This fixes that be decrementing the use count
of dependencies in all the locations where a module is
unloaded, and shutdown now finishes cleanly again.