auth.c: Add login caching to improve performance.
This adds some simple login caching to the BBS
to improve performance. This is a double-edged
sword obviously, so several things were considered
in the implementation.

The reason for adding a caching layer is because
many of the protocols provided by the BBS require
frequent authentication, e.g. HTTP, IMAP, etc.
In such cases, waiting for secure, slow hashes
like bcrypt causes a perceivable lag that users
can notice. These properties are desirable for
security, but not for performance.

As a tradeoff, the SHA256 hashes of passwords
are now stored in memory for recent logins. If
the same user logs in within an hour after such
a hash is stored, from the same IP address, then
as long as the hashes match, the login is granted.
This limits the scope of any possible attacks against
this mechanism.

net_http previously had a caching mechanism very
similar to this. Because that functionality is now
redundant, as bbs_authenticate handles this internally,
this module-specific cache has been removed.